+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- BREAKING CHANGE: Remove support for Terraform Google provider version `< 4.0`.
+
 ## [0.0.12]
 
 ### Changed